fix(be): Goal 엔티티 및 dto의 금액 자료형 long으로 수정 (#131)

+++ backend/src/main/java/com/back/domain/goal/dto/GoalDto.java
@@ -9,8 +9,8 @@ public record GoalDto(
         int id,
         int memberId,
         String description,
-        int currentAmount,
-        int targetAmount,
+        long currentAmount,
+        long targetAmount,
         LocalDateTime deadline,
         GoalStatus status
 ) {
